Move database function. Delete story states as well.
